[FIX] level is not a float, but an integer (lp:c2c-addons/6.1  rev 89.1.14)

+++ account_credit_control/line.py
@@ -100,7 +100,7 @@ class CreditControlLine(Model):
 
         'level': fields.related('policy_level_id',
                                 'level',
-                                type='float',
+                                type='integer',
                                 relation='credit.control.policy',
                                 string='Level',
                                 store=True,
